Move setupOptions to setupActions file and update order of imports
