+++ src/components/SidePanel/StartStop.jsx
@@ -41,8 +41,8 @@ import { useDispatch, useSelector } from 'react-redux';
 import { unit } from 'mathjs';
 import { Group } from 'pc-nrfconnect-shared';
 
-import { setupOptions } from '../../actions/deviceActions';
 import { samplingStart, samplingStop } from '../../actions/samplingActions';
+import setupOptions from '../../actions/setupActions';
 import { Slider } from '../../from_pc-nrfconnect-shared';
 import { appState } from '../../reducers/appReducer';
 import {
